# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerResponse": {
        "description": "Evidence-grounded answer contract.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "confidence": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "degraded_mode": {
            "title": "Degraded Mode",
            "type": "boolean"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "retrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "retrieval_status": {
            "$ref": "#/components/schemas/RetrievalStatus",
            "default": "resolved"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "confidence",
          "evidence",
          "warnings",
          "degraded_mode"
        ],
        "title": "AnswerResponse",
        "type": "object"
      },
...
      "ErrorPayload": {
        "description": "Stable, human-safe API error information.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "suggested_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Action"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id",
          "retryable"
        ],
        "title": "ErrorPayload",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Top-level error envelope shared by every API failure.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvidenceResponse": {
        "description": "One validated transcript citation returned with an answer.",
        "properties": {
          "cue_ids": {
            "items": {
              "type": "string"
            },
            "title": "Cue Ids",
            "type": "array"
          },
          "end_ms": {
            "exclusiveMinimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "modality": {
            "const": "transcript",
            "default": "transcript",
            "title": "Modality",
            "type": "string"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "retrieval_unit_id": {
            "title": "Retrieval Unit Id",
            "type": "string"
          },
          "start_ms": {
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          },
          "video_id": {
            "format": "uuid",
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "retrieval_unit_id",
          "cue_ids",
          "quote",
          "start_ms",
          "end_ms"
        ],
        "title": "EvidenceResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportVideoRequest": {
        "description": "A caller-supplied video locator accepted by the source registry.",
        "properties": {
          "source_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "source_url"
        ],
        "title": "ImportVideoRequest",
        "type": "object"
      },
      "ImportVideoResponse": {
        "description": "Prompt durable-import acknowledgement with idempotent job reuse.",
        "properties": {
          "job": {
            "$ref": "#/components/schemas/IngestionJobResponse"
          },
          "reused": {
            "title": "Reused",
            "type": "boolean"
          }
        },
        "required": [
          "job",
          "reused"
        ],
        "title": "ImportVideoResponse",
        "type": "object"
      },
...
      "IngestionJobResponse": {
        "description": "Current durable job projection with source identity and lifecycle timestamps.",
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "cancel_requested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Requested At"
          },
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "input_fingerprint": {
            "title": "Input Fingerprint",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "last_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Code"
          },
          "last_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Message"
          },
          "last_error_retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Retryable"
          },
          "source_kind": {
            "$ref": "#/components/schemas/VideoSourceKind"
          },
          "stage": {
            "$ref": "#/components/schemas/IngestionStage"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/IngestionJobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "video_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Id"
          }
        },
        "required": [
          "job_id",
          "source_kind",
          "external_id",
          "canonical_url",
          "input_fingerprint",
          "status",
          "stage",
          "attempt",
          "created_at",
          "updated_at"
        ],
        "title": "IngestionJobResponse",
        "type": "object"
      },
...
      "IngestionStage": {
        "description": "Phase 2 stages in execution order, excluding later multimodal capabilities.",
        "enum": [
          "source_resolution",
          "metadata",
          "caption_retrieval",
          "audio_acquisition",
          "transcription",
          "chunking",
          "persistence",
          "embedding",
          "indexing",
          "cleanup",
          "completed"
        ],
        "title": "IngestionStage",
        "type": "string"
      },
      "JsonValue": {},
...
      "QuestionRequest": {
        "additionalProperties": false,
        "description": "A non-empty transcript question.",
        "properties": {
          "allow_fallback": {
            "default": true,
            "title": "Allow Fallback",
            "type": "boolean"
          },
          "limit": {
            "default": 8,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "mode": {
            "$ref": "#/components/schemas/RetrievalMode",
            "default": "reranked"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "selected_anchor": {
            "anyOf": [
              {
                "pattern": "^[a-f0-9]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Anchor"
          }
        },
        "required": [
          "question"
        ],
        "title": "QuestionRequest",
        "type": "object"
      },
...
      "RetrievalMode": {
        "enum": [
          "lexical",
          "dense",
          "hybrid",
          "reranked"
        ],
        "title": "RetrievalMode",
        "type": "string"
      },
      "RetrievalStatus": {
        "enum": [
          "resolved",
          "anchor_selection_required",
          "anchor_unresolved"
        ],
        "title": "RetrievalStatus",
        "type": "string"
      },
...
      "TranscriptResponse": {
        "description": "Complete transcript projection used by the Phase 1 UI.",
        "properties": {
          "cues": {
            "items": {
              "$ref": "#/components/schemas/TranscriptCueResponse"
            },
            "title": "Cues",
            "type": "array"
          },
          "retrieval_units": {
            "items": {
              "$ref": "#/components/schemas/RetrievalUnitResponse"
            },
            "title": "Retrieval Units",
            "type": "array"
          },
          "transcription": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptionRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "video": {
            "$ref": "#/components/schemas/VideoResponse"
          }
        },
        "required": [
          "video",
          "cues",
          "retrieval_units",
          "transcription"
        ],
        "title": "TranscriptResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VideoResponse": {
        "description": "Safe canonical metadata returned by import and detail endpoints.",
        "properties": {
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "channel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Name"
          },
          "duration_ms": {
            "exclusiveMinimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "index_ready": {
            "title": "Index Ready",
            "type": "boolean"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transcript_ready": {
            "default": true,
            "title": "Transcript Ready",
            "type": "boolean"
          },
          "video_id": {
            "format": "uuid",
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "source_kind",
          "external_id",
          "canonical_url",
          "title",
          "duration_ms",
          "index_ready"
        ],
        "title": "VideoResponse",
        "type": "object"
      },
      "VideoSourceKind": {
        "description": "Supported source families for the transcript-first slice.",
        "enum": [
          "youtube",
          "local_file"
        ],
        "title": "VideoSourceKind",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Galaxy Frog API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/videos/import": {
      "post": {
        "description": "Create or reuse a durable job without running provider work in the request.",
        "operationId": "import_video_v1_videos_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportVideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportVideoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Import Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/v1/videos/{video_id}": {
      "get": {
        "description": "Return safe canonical metadata for one imported video.",
        "operationId": "get_video_v1_videos__video_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/v1/videos/{video_id}/questions": {
      "post": {
        "description": "Retrieve transcript evidence and return a deterministically validated answer.",
        "operationId": "ask_question_v1_videos__video_id__questions_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ask Question",
        "tags": [
          "videos"
        ]
      }
    },
...
    "/v1/videos/{video_id}/transcript": {
      "get": {
        "description": "Return exact cues and retrieval units with complete provenance.",
        "operationId": "get_transcript_v1_videos__video_id__transcript_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Transcript",
        "tags": [
          "videos"
        ]
      }
    }
  }
}